CLI-982 Merge CliContext types into a single cli-context module
