+++ CLAUDE.md
@@ -38,7 +38,7 @@ Each command lives in `src/commands/`. The command tree is built by `createComma
 
 To add a new command: add it to `createCommandTree()` in `src/commands/command-tree.ts` and implement the logic in a new folder under `src/commands/`.
 Please declare commands using the type defined in `src/commands/sonar-command.ts`.
-By default, new commands should register a `authenticatedAction()`; the handler's first argument is a `CliAuthenticatedContext` (`auth`, plus resolved `isAlpha` / `isBeta` for this command's stage — see `src/commands/cli-authenticated-context.ts`). `anonymousAction()` receives a `CliContext` parent (`isAlpha` / `isBeta` only — see `src/commands/cli-context.ts`). Only technical commands use `anonymousAction()`.
+By default, new commands should register a `authenticatedAction()`; the handler's first argument is a `CliAuthenticatedContext` (`auth`, plus resolved `isAlpha` / `isBeta` for this command's stage — see `src/commands/cli-context.ts`). `anonymousAction()` receives a `CliContext` parent (`isAlpha` / `isBeta` only — see `src/commands/cli-context.ts`). Only technical commands use `anonymousAction()`.
 Mark lifecycle with `.stage(Stage.Stable)` (default), `.stage(Stage.Alpha)`, `.stage(Stage.Beta())` (Open Beta), or `.stage(Stage.Beta('cli.beta.flag-key'))` (Private Beta, gated by LaunchDarkly). Visibility is decided via `CliRuntime` (`isAlphaEnabled` from `SONARQUBE_CLI_ALPHA`, `isPrivateBetaEnabled` from LaunchDarkly): `.stage()` keeps or detaches the command after Commander’s eager `.command()` attach. `createCommandTree` is async: it probes for Private Beta flag keys, then only when keys exist calls optional `loadPrivateBetaContext(flagKeys)` (wired in `src/index.ts` to resolve auth + `resolvePrivateBetaFlags`, 2-hour cache under `~/.sonar/sonarqube-cli/launch-darkly/`) and rebuilds the tree so Private Beta commands register only when the flag is `true`. With no keys (or no loader), LaunchDarkly is never contacted; docs call `await createCommandTree()` without a loader so Private Beta stays omitted. Set `SONARQUBE_CLI_LAUNCHDARKLY_ENV=dev` to use the LaunchDarkly `dev` environment (default `production`). Open Beta appears in public docs; Private Beta and Alpha do not.
 Visible top-level commands use their main `description()` as root help copy. Every visible top-level command should declare `rootHelp({ category })` so the custom root menu can group commands with blank lines between the `core`, `data`, `integrate`, and `cli-management` sections. The order within each group follows the declaration order in `src/commands/command-tree.ts`, so add or move top-level commands in the desired display order there. When a visible top-level command has visible subcommands, the custom root menu derives its label automatically as `name <sub1|sub2|...>` unless that command also sets `rootHelp({ expandSubcommands: true })`, in which case the parent command is rendered as just `name` and its visible subcommands are listed individually below it. Add `rootHelp({ label })` only when the default label is not enough. If a command should disappear from help entirely, declare it with Commander `{ hidden: true }` and leave an inline comment when hiding a public compatibility command.
 The bare `sonar analyze` command accepts `-p, --project <project>` for the agentic portion of the combined flow. Use it to override auto-detection or to run `sonar analyze` without an installed project integration; `--branch` remains specific to `analyze agentic` / `verify`. SQAA accepts `--depth STANDARD|DEEP`: change-set and multi `--file` default to DEEP; single `--file` defaults to STANDARD unless `--depth DEEP` is set. Per-edit hooks force STANDARD; end-of-turn instructions tell agents to always pass `--depth DEEP` (even for a single file).
