--- src/commands/cli-authenticated-context.ts
@@ -1,39 +1,0 @@
-/*
- * SonarQube CLI
- * Copyright (C) SonarSource Sàrl
- * mailto:info AT sonarsource DOT com
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 3 of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
- */
-
-import type { ResolvedAuth } from '@/core/auth/auth-resolver.ts';
-
-import { CliContext } from './cli-context.ts';
-
-/**
- * Per-command invocation context for authenticated handlers.
- *
- * Built by `SonarCommand.authenticatedAction`. Extends {@link CliContext} with
- * resolved auth for this invocation.
- */
-export class CliAuthenticatedContext extends CliContext {
-  constructor(
-    readonly auth: ResolvedAuth,
-    isAlpha: boolean,
-    isBeta: boolean,
-  ) {
-    super(isAlpha, isBeta);
-  }
-}
