+++ src/commands/integrate/integrate-bare.ts
@@ -18,7 +18,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-import type { CliAuthenticatedContext } from '@/commands/cli-authenticated-context.ts';
+import type { CliAuthenticatedContext } from '@/commands/cli-context.ts';
 import { CommandFailedError } from '@/core/command-error.ts';
 import { selectPrompt } from '@/core/ui';
